State of this project?
Does anyone know what the state of this project is? The last updates dates back 6 years. There are quite a few forks, but none seem active.  
There was a PR from 2020 (https://github.com/jzdziarski/mod_evasive/pull/3) but that was closed without merge.  

Can we still considder the project as stable? It's still installable from the standard apt repositories on Debian, but given that the README 
in the repo still mentions Apache 1.3 and 2.0, but doesn't mention 2.4, doesn't quite feel right.